Increase docker API timeout
While we started testing parallel usage we discovered that the 5 seconds
timeout was too small.

Increating this to 10s to assure it does not fail. In the future we may
want to make it configurable or to default to 60 seconds which is the
current default on python docker api client.