Add option to add markers for unavailable drivers
This will allow users to run only scenarios that can be run on their
machines.

Initial implementation does validation only for delegated and
docker drivers.

For delegated driver, we assume current inventory is valid if we
find presence of zuul variable. This should protect againt accidental
execution of ansible tasks on localhost.