Add base config option
Up until now, it was fairly inconvenient to run molecule test suites
that used based configuration file to reduce the duplication of
content in the scenario configs.

This commit adds a command line flag and related INI config option
that allows us to pass --base-config flag to molecule when running
tests.