Make driver and platforms config sections optional (#87)
Current implementation mandates that at least driver and platfroms
sections are present in all molecule.yml scenario configuration files.
But because those two pieces of information can also come from the
base configuration file, it does not feel right to fail when they are
not present in the scenario configuration.

This commit makes those two sections optional. Missing platfroms
section is simply interpreted as an empty list since this does not
change the final outcome at all.

Missing driver section is treated a bit differently because the driver
name is used in a few different places. In order not to break
backward-compatibility, we use the "no_driver" string as driver name if
the driver section is not present or if the driver name is not set in
the scenario config.